requirements.txt
numpy==1.22.4
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.10.1
pyqtgraph==0.12.4
pyserial==3.5

certifi==2021.10.8
charset-normalizer==2.0.12
cx-Freeze==6.10
cx-Logging==3.0
digi-xbee==1.4.1
gmplot==1.4.1
idna==3.3
importlib-metadata==4.11.4
lief==0.12.1
pandas==1.4.2
PyMeasure==0.10.0
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.10.1
pyqtgraph==0.12.4
pyserial==3.5
python-dateutil==2.8.2
pytz==2022.1
PyVISA==1.11.3
requests==2.27.1
six==1.16.0
typing-extensions==4.1.1
urllib3==1.26.9
zipp==3.8.0

